src/requirements.txt
pandas
requests
boto3
beautifulsoup4
pyarrow
python-dotenv